fix(psalm): drain all errors introduced by SQL-DTO-AUDIT commit
- UploadService: drop redundant (int) casts on getimagesize result
  (already typed as int by Psalm)
- AlbumController: null-coalesce id when building edit URL
  (CreateCategoryResult::id is int|null; non-null in the else branch
  but Psalm can't infer that from the success factory)
- MaintenanceController: pass $formats_to_delete directly to
  deleteByFormatIds — values are already int, map+cast was redundant
- PhotoController/PictureController: explicit (float) cast before
  dividing int by float literal (Psalm strict binary operands mode)
- PictureController: change `if ($slideshowActive)` to
  `if ($slideshowActive && $slideshow_params !== null)` so Psalm
  narrows the type inside the block without needing assert()
- PictureService::encodeSlideshowParams: cast period to string before
  concatenation (period is int|float)
- UserService: cast $error to string before passing to
  UpdateUserResult::error (validateMailAddress may return null)

+++ src/Piwigo/Controller/Admin/PhotoController.php
@@ -558,7 +558,7 @@ final class PhotoController implements AdminSubControllerInterface
                 'ext'          => $fmt->ext,
                 'download_url' => $this->urlGenerator->actionFormat($fmt->formatId),
                 'label'        => Lang::has($lang_key) ? Lang::t($lang_key) : $label,
-                'filesize'     => round(($fmt->filesize ?? 0) / 1024.0, 2),
+                'filesize'     => round((float) ($fmt->filesize ?? 0) / 1024.0, 2),
             ];
         }, $this->imageFormatRepository->findByImageIds([$picFmtIdInt]));
